Add -dry-run
Quoting https://github.com/rfjakob/cshatag/issues/22 :

 To my understanding, if either the mtime or the checksum of a file (or
 both) have changed, "the status of the file is printed to stdout and
 the stored checksum is updated" (this last part is taken directly from
 the readme). This means that if something happened to your file that
 you weren't aware of (for example, if you edited it by mistake), you'll
 only find out when running `cshatag` the first time. The second time,
 the checksum will already have been updated and everything looks
 normal. I can foresee many situations where this might be a problem,
 for example if your computer loses power before you have a chance to
 inspect the output or even if you forget that you've run the program
 and run it again.

 In [rsync](https://github.com/WayneD/rsync), you can dry run the
 program using the `-n` flag. This means that you get to see all error
 messages, warnings, et cetera, without actually executing the core
 function of the program, namely to copy files. Would it be desirable to
 have something similar implemented in `cshatag`?

Closes https://github.com/rfjakob/cshatag/issues/22
