+++ check.go
@@ -3,11 +3,9 @@ package main
 import (
 	"bytes"
 	"crypto/sha256"
-	"errors"
 	"fmt"
 	"io"
 	"os"
-	"runtime"
 	"strconv"
 	"strings"
 	"syscall"
@@ -26,7 +24,7 @@ type fileTimestamp struct {
 
 func zeroFileTimeStamp() fileTimestamp {
 	return fileTimestamp{
-		s: uint64(0),
+		s:  uint64(0),
 		ns: uint32(0),
 	}
 }
@@ -99,46 +97,6 @@ func getActualAttr(f *os.File) (attr fileAttr, err error) {
 	}
 	attr.sha256 = []byte(fmt.Sprintf("%x", h.Sum(nil)))
 	return attr, nil
-}
-
-// storeAttr stores "attr" into extended attributes.
-// Should look like this afterwards:
-//
-//     $ getfattr -d foo.txt
-//     user.shatag.sha256="dc9fe2260fd6748b29532be0ca2750a50f9eca82046b15497f127eba6dda90e8"
-//     user.shatag.ts="1560177334.020775051"
-func storeAttr(f *os.File, attr fileAttr) (err error) {
-	if runtime.GOOS == "darwin" {
-		// SMB or MacOS bug: when working on an SMB mounted filesystem on a Mac, it seems the call
-		// to `fsetxattr` does not update the xattr but removes it instead. So it takes two runs
-		// of `cshatag` to update the attribute.
-		// To work around this issue, we remove the xattr explicitly before setting it again.
-		// https://github.com/rfjakob/cshatag/issues/8
-		removeAttr(f)
-	}
-	err = xattr.FSet(f, xattrTs, []byte(attr.ts.prettyPrint()))
-	if err != nil {
-		return
-	}
-	err = xattr.FSet(f, xattrSha256, attr.sha256)
-	return
-}
-
-// removeAttr removes any previously stored extended attributes. Returns an error
-// if removal of either the timestamp or checksum xattrs fails.
-func removeAttr(f *os.File) error {
-	err1 := xattr.FRemove(f, xattrTs)
-	err2 := xattr.FRemove(f, xattrSha256)
-	if err1 != nil && err2 != nil {
-		return errors.New(err1.Error() + "  " + err2.Error())
-	}
-	if err1 != nil {
-		return err1
-	}
-	if err2 != nil {
-		return err2
-	}
-	return nil
 }
 
 // printComparison prints something like this:
