+++ cshatag.1
@@ -47,6 +47,8 @@ rewritten in Go in 2019.
 
 .SH OPTIONS
 
+-dry-run    don't make any changes
+.br
 -recursive  recursively process the contents of directories
 .br
 -remove     remove cshatag's xattrs from FILE
