+++ tests/11.expected
@@ -1,0 +1,1 @@
+<removed xattr> foo.txt
