+++ tests/9.expected
@@ -1,0 +1,3 @@
+<new> foo.txt
+ stored: 0000000000000000000000000000000000000000000000000000000000000000 0000000000.000000000
+ actual: e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855 1546297200.000000000
